+++ src/lib/layout/coverTitle.svelte
@@ -5,12 +5,14 @@
 </script>
 
 <Heading size="4" tag="h1" trimmed={false}>
-    {#if href}
-        <a class="button is-text is-only-icon" {href} aria-label="page back">
-            <span class="icon-cheveron-left" aria-hidden="true" />
-        </a>
-    {/if}
-    <span class="text u-trim-1">
-        <slot />
-    </span>
+    <div class="u-flex u-cross-center">
+        {#if href}
+            <a class="button is-text is-only-icon" {href} aria-label="page back">
+                <span class="icon-cheveron-left" aria-hidden="true" />
+            </a>
+        {/if}
+        <span class="text u-trim-1">
+            <slot />
+        </span>
+    </div>
 </Heading>
