+++ src/routes/console/project-[project]/databases/database-[database]/createCollection.svelte
